+++ src/storage/file_storage.rs
@@ -1,118 +1,10 @@
-use std::io::Read;
-use std::io::Seek;
-use std::io::Write;
-
-#[allow(unused_imports)]
-use super::serialize::Serialize;
-use super::storage_impl::StorageImpl;
-use super::storage_record::StorageRecord;
-use super::storage_record_with_index::StorageRecordWithIndex;
+use super::file_storage_data::FileStorageData;
 use super::storage_records::StorageRecords;
 use super::write_ahead_log::WriteAheadLog;
-use super::write_ahead_log_record::WriteAheadLogRecord;
 use super::Storage;
 use crate::db_error::DbError;
 
-pub(crate) struct FileStorage {
-    file: std::fs::File,
-    #[allow(dead_code)]
-    filename: String,
-    records: StorageRecords,
-    wal: WriteAheadLog,
-    wal_filename: String,
-    transactions: u64,
-}
-
-impl StorageImpl for FileStorage {
-    fn begin_transaction(&mut self) {
-        self.transactions += 1;
-    }
-
-    fn clear_wal(&mut self) -> Result<(), DbError> {
-        self.wal.clear()
-    }
-
-    fn create_index(&mut self, position: u64, size: u64) -> i64 {
-        self.records.create(position, size)
-    }
-
-    fn end_transaction(&mut self) -> bool {
-        if self.transactions != 0 {
-            self.transactions -= 1;
-        }
-
-        self.transactions == 0
-    }
-
-    fn indexes_by_position(&self) -> Vec<i64> {
-        self.records.indexes_by_position()
-    }
-
-    fn insert_wal_record(&mut self, record: WriteAheadLogRecord) -> Result<(), DbError> {
-        self.wal.insert(record)
-    }
-
-    fn read_exact(&mut self, buffer: &mut Vec<u8>) -> Result<(), DbError> {
-        Ok(self.file.read_exact(buffer)?)
-    }
-
-    fn record(&self, index: i64) -> Result<StorageRecord, DbError> {
-        Ok(self
-            .records
-            .get(index)
-            .ok_or_else(|| DbError::from(format!("index '{}' not found", index)))?
-            .clone())
-    }
-
-    fn record_mut(&mut self, index: i64) -> &mut StorageRecord {
-        self.records
-            .get_mut(index)
-            .expect("validated by previous call to FileStorage::record()")
-    }
-
-    fn remove_index(&mut self, index: i64) {
-        self.records.remove(index);
-    }
-
-    fn seek(&mut self, position: std::io::SeekFrom) -> Result<u64, DbError> {
-        Ok(self.file.seek(position)?)
-    }
-
-    fn set_len(&mut self, len: u64) -> Result<(), DbError> {
-        Ok(self.file.set_len(len)?)
-    }
-
-    fn set_records(&mut self, records: Vec<StorageRecordWithIndex>) {
-        self.records = StorageRecords::from(records);
-    }
-
-    fn wal_records(&mut self) -> Result<Vec<WriteAheadLogRecord>, DbError> {
-        self.wal.records()
-    }
-
-    fn write_all(&mut self, bytes: &[u8]) -> Result<(), DbError> {
-        Ok(self.file.write_all(bytes)?)
-    }
-}
-
-impl Storage for FileStorage {}
-
-impl Drop for FileStorage {
-    fn drop(&mut self) {
-        if self.apply_wal().is_ok() {
-            let _ignore1 = self.wal.clear();
-            let _ignore2 = std::fs::remove_file(&self.wal_filename);
-        }
-    }
-}
-
-impl TryFrom<&str> for FileStorage {
-    type Error = DbError;
-
-    fn try_from(filename: &str) -> Result<Self, Self::Error> {
-        Self::try_from(filename.to_string())
-    }
-}
+pub(crate) type FileStorage = Storage<FileStorageData>;
 
 fn wal_filename(filename: &str) -> String {
     let pos;
@@ -137,16 +29,18 @@ impl TryFrom<String> for FileStorage {
         let wal_filename = wal_filename(&filename);
 
         let mut storage = FileStorage {
-            file: std::fs::OpenOptions::new()
-                .write(true)
-                .create(true)
-                .read(true)
-                .open(&filename)?,
-            filename,
-            records: StorageRecords::default(),
-            wal: WriteAheadLog::try_from(&wal_filename)?,
-            wal_filename,
-            transactions: 0,
+            data: FileStorageData {
+                file: std::fs::OpenOptions::new()
+                    .write(true)
+                    .create(true)
+                    .read(true)
+                    .open(&filename)?,
+                filename,
+                records: StorageRecords::default(),
+                wal: WriteAheadLog::try_from(&wal_filename)?,
+                wal_filename,
+                transactions: 0,
+            },
         };
 
         storage.apply_wal()?;
@@ -156,9 +50,19 @@ impl TryFrom<String> for FileStorage {
     }
 }
 
+impl TryFrom<&str> for FileStorage {
+    type Error = DbError;
+
+    fn try_from(filename: &str) -> Result<Self, Self::Error> {
+        Self::try_from(filename.to_string())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::storage::storage_record::StorageRecord;
+    use crate::storage::Serialize;
     use crate::test_utilities::test_file::TestFile;
 
     #[test]
